% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologically_relevant_lncRNAs.R
\name{biologically_relevant_lncRNAs}
\alias{biologically_relevant_lncRNAs}
\title{Extract the biologically relevant lncRNAs
This function extracts the most biologically relevant lncRNAs from a list of highly expressed lncRNAs only detected by Kallisto.}
\usage{
biologically_relevant_lncRNAs(
  candidates,
  ratio_threshold,
  CR_threshold,
  SI_threshold
)
}
\arguments{
\item{candidates}{List of highly expressed lncRNAs only detected by Kallisto}

\item{ratio_threshold}{Threshold to remove lncRNAs whose ratio of expression between Kallisto and CellRanger is smaller than this defined threshold.}

\item{CR_threshold}{Threshold to remove lncRNAs that have a CellRanger expression higher than this defined threshold.}

\item{SI_threshold}{Threshold to remove lncRNAs whose SI is smaller than this defined threshold.}
}
\value{
most biologically relevant lncRNAs
}
\description{
Extract the biologically relevant lncRNAs
This function extracts the most biologically relevant lncRNAs from a list of highly expressed lncRNAs only detected by Kallisto.
}
