% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr_data.R
\name{crispr_info}
\alias{crispr_info}
\title{This function checks if highly-expressed exclusive lncRNAs have been functionally validated with external CRISPRi studies}
\usage{
crispr_info(crispr_data, candidates_lncRNAs)
}
\arguments{
\item{crispr_data}{Public CRISPRi data with lncRNAs that have been externally validated}

\item{candidates_lncRNAs}{Highly-expressed exclusive lncRNAS}
}
\value{
Boolean vector on whether or not the highly-expressed exclusive lncRNAs has a hit in public CRISPRi studies
}
\description{
This function checks if highly-expressed exclusive lncRNAs have been functionally validated with external CRISPRi studies
}
