% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARKAS.R, R/ELATUS.R
\name{ELATUS}
\alias{ELATUS}
\title{ELATUS workflow
This workflow loads the Kallisto and CellRanger raw count matrices, performs emptydrops, then doublet removal, a quality control filtering. Next it gets the highly expressed lncRNAs only detected by Kallisto and the ratio of their expression between Kallisto/CellRanger and finally, after normalization and clustering it calculates the Specificity Index of each of them. The last step is generating the list of biologically relevant lncRNAs.}
\usage{
ELATUS(
  kallisto_path,
  kallisto_name,
  cellRanger_path,
  organism,
  lower_emptydrops,
  EmptyDrops_FDR_thres,
  cells_mito_threshold,
  cells_max_threshold,
  cells_min_genes_detected_threshold,
  threshold_minumun_gene_counts,
  threshold_cells_detected,
  dimred_clustering,
  k_neighbors,
  ratio_threshold,
  CR_threshold,
  SI_threshold
)

ELATUS(
  kallisto_path,
  kallisto_name,
  cellRanger_path,
  organism,
  lower_emptydrops,
  EmptyDrops_FDR_thres,
  cells_mito_threshold,
  cells_max_threshold,
  cells_min_genes_detected_threshold,
  threshold_minumun_gene_counts,
  threshold_cells_detected,
  dimred_clustering,
  k_neighbors,
  ratio_threshold,
  CR_threshold,
  SI_threshold
)
}
\arguments{
\item{kallisto_path}{Path to the Kallisto raw count matrix (As an example: kallisto_path=system.file("extdata", "kallisto_example_raw_matrix", package = "ARKAS"))}

\item{kallisto_name}{Name of the Kallisto raw count matrix (As an example kallisto_name="cells_genes_NO_multimapping")}

\item{cellRanger_path}{Path to the CellRanger raw count matrix (As an example :cellRanger_path=system.file("extdata", "cellRanger_example_raw_matrix", package = "ARKAS"))}

\item{organism}{Human or Mouse dataset (for the example "Human")}

\item{lower_emptydrops}{A numeric scalar specifying the lower bound on the total UMI count, at or below which all barcodes are assumed to correspond to empty droplets. For the example, 1000}

\item{EmptyDrops_FDR_thres}{FDR threshold to classifly empty droplets (For the example, 0.01)}

\item{cells_mito_threshold}{Mitochondrial content (\%) threshold. Keep cells with less mitocondrial content than the defined threshold (For the example = 15)}

\item{cells_max_threshold}{Keep cells with less counts than the defined threshold (For the example = 30000)}

\item{cells_min_genes_detected_threshold}{Keep cells with counts in more genes than the defined threhold (For the example = 500)}

\item{threshold_minumun_gene_counts}{Select genes with more than this total counts  (For the paper = 250 counts)}

\item{threshold_cells_detected}{Select genes present in at least a number of cells higher than this threshold (For the paper = 25)}

\item{dimred_clustering}{dimensionality reduction (For the example "PCA")}

\item{k_neighbors}{the number of nearest neighbors used to construct the graph. Choose a smaller k to more but smaller clusters as lncRNAs tend to be expressed in small subpopulations. (in this example, k=5)}

\item{ratio_threshold}{Threshold to remove lncRNAs whose ratio of expression between Kallisto and CellRanger is smaller than this defined threshold (For the paper = 40)}

\item{CR_threshold}{Threshold to remove lncRNAs that have a CellRanger expression higher than this defined threshold (For the paper = 10)}

\item{SI_threshold}{Threshold to remove lncRNAs whose SI is smaller than this defined threshold (For the paper = 0.15)}
}
\value{
A list with most biologically relevant lncRNAs

A list with most biologically relevant lncRNAs
}
\description{
ELATUS workflow
This workflow loads the Kallisto and CellRanger raw count matrices, performs emptydrops, then doublet removal, a quality control filtering. Next it gets the highly expressed lncRNAs only detected by Kallisto and the ratio of their expression between Kallisto/CellRanger and finally, after normalization and clustering it calculates the Specificity Index of each of them. The last step is generating the list of biologically relevant lncRNAs.

ELATUS workflow
This workflow loads the Kallisto and CellRanger raw count matrices, performs emptydrops, then doublet removal, a quality control filtering. Next it gets the highly expressed lncRNAs only detected by Kallisto and the ratio of their expression between Kallisto/CellRanger and finally, after normalization and clustering it calculates the Specificity Index of each of them. The last step is generating the list of biologically relevant lncRNAs.
}
