% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELATUS_filtered.r
\name{ELATUS_filtered}
\alias{ELATUS_filtered}
\title{ELATUS workflow
ELATUS workflow but working directly with the filtered objects for both Cell Ranger and Kallisto}
\usage{
ELATUS_filtered(
  organism,
  kallisto_sce,
  cellRanger_sce,
  gene_names,
  threshold_minumun_gene_counts,
  threshold_cells_detected,
  dimred_clustering,
  k_neighbors,
  ratio_threshold,
  CR_threshold,
  SI_threshold
)
}
\arguments{
\item{organism}{Human or Mouse dataset (for the example "Mouse")}

\item{kallisto_sce}{Kallisto SingleCellExperiment object}

\item{gene_names}{Are gene names the rownames of the SingleCellExperiment object}

\item{threshold_minumun_gene_counts}{Select genes with more than this total counts  (For the paper = 250 counts)}

\item{threshold_cells_detected}{Select genes present in at least a number of cells higher than this threshold (For the paper = 25)}

\item{dimred_clustering}{dimensionality reduction (For the example "PCA")}

\item{k_neighbors}{the number of nearest neighbors used to construct the graph. Choose a smaller k to more but smaller clusters as lncRNAs tend to be expressed in small subpopulations. (in this example, k=5)}

\item{ratio_threshold}{Threshold to remove lncRNAs whose ratio of expression between Kallisto and CellRanger is smaller than this defined threshold (For the paper = 40)}

\item{CR_threshold}{Threshold to remove lncRNAs that have a CellRanger expression higher than this defined threshold (For the paper = 10)}

\item{SI_threshold}{Threshold to remove lncRNAs whose SI is smaller than this defined threshold (For the paper = 0.15)}

\item{cellRanger_path}{CellRanger SingleCellExperiment object}
}
\value{
A list with most biologically relevant lncRNAs
}
\description{
ELATUS workflow
ELATUS workflow but working directly with the filtered objects for both Cell Ranger and Kallisto
}
