% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELATUS_pre_mRNA.R
\name{ELATUS_premRNA}
\alias{ELATUS_premRNA}
\title{ELATUS workflow ofor snRNA-seq using a pre-mRNA reference:This workflow loads the Kallisto and CellRanger raw count matrices as well as Kallisto allowing for multimappers, performs emptydrops, then doublet removal. Next it gets the highly expressed lncRNAs only detected by both Kallisto and Kallisto multimappers and the ratio of their expression between Kallisto/CellRanger and finally, after normalization and clustering it calculates the Specificity Index of each of them. The last step is generating the list of biologically relevant lncRNAs. In addition it identifies highly-expressed lncRNAs exclusively detected by Kallisto whose functionality has been proven by external biobliography (CRISPR screenings...) and those highly-expressed lncRNAs robustly detected by both Cell Ranger and Kallisto and by Cell Ranger and Kallisto multimappers.}
\usage{
ELATUS_premRNA(
  kallisto_path,
  kallisto_name,
  cellRanger_path,
  kallisto_multimappers_path,
  kallisto_multimappers_name,
  organism,
  lower_emptydrops,
  EmptyDrops_FDR_thres,
  threshold_minumun_gene_counts,
  threshold_cells_detected,
  dimred_clustering,
  k_neighbors,
  ratio_threshold,
  CR_threshold,
  SI_threshold
)
}
\arguments{
\item{kallisto_path}{Path to the Kallisto raw count matrix (As an example: kallisto_path=/home/egonie/kike/phd/test_data/10X/MULTIOME_PBMCS_3K/01.Kallisto/output_bus_transcriptome_introns/bustools_results_no_multimappers)}

\item{kallisto_name}{Name of the Kallisto raw count matrix (As an example kallisto_name="cells_genes_NO_multimapping")}

\item{cellRanger_path}{Path to the CellRanger raw count matrix (As an example :cellRanger_path="/home/egonie/kike/phd/test_data/10X/MULTIOME_PBMCS_3K/01.CellRanger/pbmc_granulocyte_sorted_3k_alldata_cellRanger_modified_GEX_test2/outs/raw_feature_bc_matrix")}

\item{kallisto_multimappers_path}{Path to the Kallisto raw count matrix (As an example: kallisto_multimappers_path="/home/egonie/kike/phd/test_data/10X/MULTIOME_PBMCS_3K/01.Kallisto/output_bus_transcriptome_introns/bustools_multimappers")}

\item{kallisto_multimappers_name}{Name of the Kallisto raw count matrix (As an example kallisto_multimappers_name="cells_genes_multimapping")}

\item{organism}{Human or Mouse dataset (for the example "Human")}

\item{lower_emptydrops}{A numeric scalar specifying the lower bound on the total UMI count, at or below which all barcodes are assumed to correspond to empty droplets. For the example, 1000}

\item{EmptyDrops_FDR_thres}{FDR threshold to classifly empty droplets (For the example, 0.01)}

\item{threshold_minumun_gene_counts}{Select genes with more than this total counts  (For the paper = 250 counts)}

\item{threshold_cells_detected}{Select genes present in at least a number of cells higher than this threshold (For the paper = 25)}

\item{dimred_clustering}{dimensionality reduction (For the example "PCA")}

\item{k_neighbors}{the number of nearest neighbors used to construct the graph. Choose a smaller k to more but smaller clusters as lncRNAs tend to be expressed in small subpopulations. (in this example, k=5)}

\item{ratio_threshold}{Threshold to remove lncRNAs whose ratio of expression between Kallisto and CellRanger is smaller than this defined threshold (For the paper = 20)}

\item{CR_threshold}{Threshold to remove lncRNAs that have a CellRanger expression higher than this defined threshold (For the paper = 20)}

\item{SI_threshold}{Threshold to remove lncRNAs whose SI is smaller than this defined threshold (For the paper = 0.15)}
}
\value{
A list with most biologically relevant lncRNAs
}
\description{
ELATUS workflow ofor snRNA-seq using a pre-mRNA reference:This workflow loads the Kallisto and CellRanger raw count matrices as well as Kallisto allowing for multimappers, performs emptydrops, then doublet removal. Next it gets the highly expressed lncRNAs only detected by both Kallisto and Kallisto multimappers and the ratio of their expression between Kallisto/CellRanger and finally, after normalization and clustering it calculates the Specificity Index of each of them. The last step is generating the list of biologically relevant lncRNAs. In addition it identifies highly-expressed lncRNAs exclusively detected by Kallisto whose functionality has been proven by external biobliography (CRISPR screenings...) and those highly-expressed lncRNAs robustly detected by both Cell Ranger and Kallisto and by Cell Ranger and Kallisto multimappers.
}
