% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptydrops_filt.r
\name{emptydrops_filt}
\alias{emptydrops_filt}
\title{Perform emptydrops filtering using https://rdrr.io/github/MarioniLab/DropletUtils/man/emptyDrops.html}
\usage{
emptydrops_filt(sce, lower, EmptyDrops_FDR_thres)
}
\arguments{
\item{sce}{SCE object}

\item{lower}{A numeric scalar specifying the lower bound on the total UMI count, at or below which all barcodes are assumed to correspond to empty droplets}

\item{EmptyDrops_FDR_thres}{FDR threshold to classifly empty droplets}
}
\value{
SCE object without empty droplets
}
\description{
Perform emptydrops filtering using https://rdrr.io/github/MarioniLab/DropletUtils/man/emptyDrops.html
}
