% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{Filtering}
\alias{Filtering}
\title{Filter SCE objects to remove cells with a lot of mitochrondrial counts, an extremely high number of counts and also cells having counts in very few genes}
\usage{
Filtering(
  sce,
  cells_mito_threshold,
  cells_max_threshold,
  cells_min_genes_detected_threshold
)
}
\arguments{
\item{cells_mito_threshold}{Mitochondrial content (\%) threshold. Keep cells with less mitocondrial content than the defined threshold.}

\item{cells_max_threshold}{Keep cells with less counts than the defined threshold.}

\item{cells_min_genes_detected_threshold}{Keep cells with counts in more genes than the defined threhold.}

\item{SCE}{SCE object to be filtered}
}
\value{
Filtered SCE object
}
\description{
Filter SCE objects to remove cells with a lot of mitochrondrial counts, an extremely high number of counts and also cells having counts in very few genes
}
