% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidates_premRNA.R
\name{get_candidates_premRNA}
\alias{get_candidates_premRNA}
\title{This function gets the highly expressed genes in snRNA-seq from each SCE and outputs those lncRNAs which are only detected by Kallisto and kallisto multimappers adding more information about them.}
\usage{
get_candidates_premRNA(
  kallisto_sce,
  cellRanger_sce,
  kallisto_multi_sce,
  threshold_minumun_gene_counts,
  threshold_cells_detected,
  lncrna_names,
  gtf,
  exclusive = T
)
}
\arguments{
\item{kallisto_sce}{Kallisto SCE object}

\item{cellRanger_sce}{CellRanger SCE object}

\item{kallisto_multi_sce}{Kallisto multimappers SCE object}

\item{threshold_minumun_gene_counts}{Select genes with more than this total counts}

\item{threshold_cells_detected}{Select genes present in at least a number of cells higher than this threshold}

\item{lncrna_names}{LncRNAs gene names}

\item{gtf}{GTF annotation}

\item{exclusive}{to get the highly-expressed lncRNAs exclusive detected with Kallisto or to get thoso commonly detected by Cell Ranger and Kallisto}
}
\value{
Candidate lncRNAs highly expressed
}
\description{
This function gets the highly expressed genes in snRNA-seq from each SCE and outputs those lncRNAs which are only detected by Kallisto and kallisto multimappers adding more information about them.
}
