% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_count_matrices.R
\name{import_CellRanger_sc}
\alias{import_CellRanger_sc}
\title{Import CellRanger raw count matrix
This function import CellRanger raw count matrix and save it in a Seurat object.}
\usage{
import_CellRanger_sc(path)
}
\arguments{
\item{path}{Path to the CellRanger raw count matrix}

\item{name}{Name of the CellRanger raw count matrix}
}
\value{
A Seurat object
}
\description{
Import CellRanger raw count matrix
This function import CellRanger raw count matrix and save it in a Seurat object.
}
