% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_count_matrices.R
\name{import_kallisto_sc}
\alias{import_kallisto_sc}
\title{Import Kallisto raw count matrix
This function import Kallisto raw count matrix and save it in a Seurat object.}
\usage{
import_kallisto_sc(path, name)
}
\arguments{
\item{path}{Path to the Kallisto raw count matrix}

\item{name}{Name of the Kallisto raw count matrix}
}
\value{
A Seurat object
}
\description{
Import Kallisto raw count matrix
This function import Kallisto raw count matrix and save it in a Seurat object.
}
