% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_doublets.R
\name{remove_doublets}
\alias{remove_doublets}
\title{Remove multiplets
This function import Kallisto raw count matrix and save it in a Seurat object.}
\usage{
remove_doublets(data)
}
\arguments{
\item{data}{SCE object already filtered with EmptyDrops}
}
\value{
SCE object without doublets
}
\description{
Remove multiplets
This function import Kallisto raw count matrix and save it in a Seurat object.
}
