% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI.R
\name{SI}
\alias{SI}
\title{Calculate the Specificity Index
This function calculates the Specificity Index from a set of genes, knowing each expression across different defined groups of cells (normally the clusters).}
\usage{
SI(sce, group_by, average_by)
}
\arguments{
\item{sce}{Kallisto SCE object}

\item{group_by}{With respect to which colData is the SI being calculated (i.e. louvain clusters)}

\item{average_by}{To define if the sumCountsAcrossCells is standardized with respect to the mean, median or nothing}
}
\value{
Candidate lncRNAs highly expressed according to Kallisto and not by CellRanger
}
\description{
Calculate the Specificity Index
This function calculates the Specificity Index from a set of genes, knowing each expression across different defined groups of cells (normally the clusters).
}
