% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidates.R
\name{get_candidates}
\alias{get_candidates}
\title{This function gets the highly expressed genes from each SCE and outputs those lncRNAs which are only detected by Kallisto adding more information about them.}
\usage{
get_candidates(
  kallisto_sce,
  cellRanger_sce,
  threshold_minumun_gene_counts,
  threshold_cells_detected,
  lncrna_names,
  gtf
)
}
\arguments{
\item{kallisto_sce}{Kallisto SCE object}

\item{cellRanger_sce}{CellRanger SCE object}

\item{threshold_minumun_gene_counts}{Select genes with more than this total counts}

\item{threshold_cells_detected}{Select genes present in at least a number of cells higher than this threshold}

\item{lncrna_names}{LncRNAs gene names}

\item{gtf}{GTF annotation}
}
\value{
Candidate lncRNAs highly expressed according to Kallisto and not by CellRanger
}
\description{
This function gets the highly expressed genes from each SCE and outputs those lncRNAs which are only detected by Kallisto adding more information about them.
}
